Tweaks to test cases for OS X.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7376 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXtackLib/src/test/net/SipMessageTest.cpp
@@ -459,7 +459,7 @@ class SipMessageTest : public CppUnit::TestCase
                     i, i);
             ASSERT_STR_EQUAL_MESSAGE(buffer, vias[i], value.data());
          }
-         CPPUNIT_ASSERT_EQUAL(i, sizeof (vias) / sizeof (vias[0]));
+         CPPUNIT_ASSERT_EQUAL(i, (unsigned int) (sizeof (vias) / sizeof (vias[0])));
 
          SipMessage testMessage2(message2, strlen(message2));
 
@@ -474,7 +474,7 @@ class SipMessageTest : public CppUnit::TestCase
                     i, i);
             ASSERT_STR_EQUAL_MESSAGE(buffer, vias[i], value.data());
          }
-         CPPUNIT_ASSERT_EQUAL(i, sizeof (vias) / sizeof (vias[0]));
+         CPPUNIT_ASSERT_EQUAL(i, (unsigned int) (sizeof (vias) / sizeof (vias[0])));
       };
